+++ com.unity.cinemachine/Editor/Editors/CinemachineConfiner2DEditor.cs
@@ -21,7 +21,7 @@ namespace Unity.Cinemachine.Editor
             this.AddMissingCmCameraHelpBox(ux);
 
             var boundsHelp = ux.AddChild(new HelpBox(
-                "Bounding Shape must be a PolygonCollider2D, BoxCollider2D, or CompositeCollider2D.",
+                "Bounding Shape 2D must be a PolygonCollider2D, BoxCollider2D, or CompositeCollider2D.",
                 HelpBoxMessageType.Warning));
             var polygonsHelp = ux.AddChild(new HelpBox(
                 "CompositeCollider2D geometry type must be Polygons.",
@@ -35,9 +35,9 @@ namespace Unity.Cinemachine.Editor
             ux.TrackAnyUserActivity(() =>
             {
                 var c = volumeProp.objectReferenceValue;
-                boundsHelp.SetVisible(c != null && c is not (PolygonCollider2D or BoxCollider2D or CompositeCollider2D));
+                boundsHelp.SetVisible(c is not (PolygonCollider2D or BoxCollider2D or CompositeCollider2D));
                 polygonsHelp.SetVisible(c is CompositeCollider2D cc && cc.geometryType != CompositeCollider2D.GeometryType.Polygons);
-                invalidCollider2D.SetVisible(c != null && Target.IsConfinerOvenNull());
+                invalidCollider2D.SetVisible(c != null && Target.BoundingShapeIsBaked && Target.IsConfinerOvenNull());
             });
 
             ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.Damping)));
@@ -80,8 +80,11 @@ namespace Unity.Cinemachine.Editor
 
             ux.Add(new Button(() =>
             {
-                Target.InvalidateBoundingShapeCache();
-                EditorUtility.SetDirty(Target);
+                for (int i = 0; i < targets.Length; i++)
+                {
+                    (targets[i] as CinemachineConfiner2D).InvalidateBoundingShapeCache();
+                    EditorUtility.SetDirty(targets[i]);
+                }
             })
             {
                 text = "Invalidate Bounding Shape Cache",